torch

+++ test_mlp/mlp_change_glass0_train&test/train_correct.py
@@ -222,7 +222,7 @@ for epoch in range(num_epochs):
     input_data_pre = torch.Tensor(torch.from_numpy(train_pre).float())
     input_data_pos = torch.Tensor(torch.from_numpy(train_pos).float())
 
-    train_label = torch.Tensor(torch.from_numpy(train_y))
+    train_label = torch.Tensor(torch.from_numpy(train_y).float()).long()
 
     out_pre = net(input_data_pre)
     out_pos = net(input_data_pos)
